+++ src/thermal.c
@@ -1,22 +1,22 @@
 #include "core.h"
 /* distribution functions */
 
-double fBose( float e ) {
+double fBose( double e ) {
   return 1. / (exp( e ) - 1.);
 };
 
-double fFermi( float e ) {
+double fFermi( double e ) {
   return 1. / (exp( e ) + 1.);
 };
 
-double f(float e, p_type X) {
+double f(double e, p_type X) {
   switch(X) {
     case F: return fFermi(e);
     case B: return fBose(e);
   }
 };
 
-double bf(float e, p_type X) {
+double bf(double e, p_type X) {
   switch(X) {
     case F: return 1.-fFermi(e);
     case B: return 1.+fBose(e);
